refactor: 컴포넌트 구조 개선
- FlexBox → foundation, Skeleton → feedBack, Button/SideBar → layout
- 모든 import 경로 수정
- GoalSelector.tsx 제거
