+++ src/app/mypage/layout.tsx
@@ -5,8 +5,8 @@ import { useRouter } from 'next/navigation';
 // import { ChevronLeft } from 'lucide-react';
 import { useAutoLogout } from '@/shared/hooks';
 // import { LogoutButton } from '@/feature/auth';
-import Button from '@/shared/components/navigation/Button';
-import { SideBar } from '@/shared/components/navigation/SideBar';
+import Button from '@/shared/components/layout/Button';
+import { SideBar } from '@/shared/components/layout/SideBar';
 
 interface HomeLayoutProps {
   children?: React.ReactNode;
