+++ src/composite/mypage/editProfile/component.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import Image from 'next/image';
-import FlexBox from '@/shared/components/layout/FlexBox';
+import FlexBox from '@/shared/components/foundation/FlexBox';
 import Badge from '@/shared/components/display/Badge';
 import Button from '@/shared/components/navigation/Button';
 import { Modal } from '@/shared/components/feedBack/Modal';
