+++ src/feature/goal/confimGoalDialogButton/component.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useState } from 'react';
 import Button from '@/shared/components/navigation/Button';
-import FlexBox from '@/shared/components/layout/FlexBox';
+import FlexBox from '@/shared/components/foundation/FlexBox';
 import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle } from '@/shared/components/dialog';
 import { useRouter } from 'next/navigation';
