+++ src/shared/components/input/Select.tsx
@@ -1,5 +1,5 @@
 import { useRef, useState, useEffect, ButtonHTMLAttributes } from 'react';
-import FlexBox from '../layout/FlexBox';
+import FlexBox from '../foundation/FlexBox';
 import { createPortal } from 'react-dom';
 
 interface SelectProps {
